BUG#27049034: SOMETIMES MEMBERS ENTER IN ERROR STATE AFTER LONG RECOVERY
In Group Replication, when a member join a group it goes through the
recovery process in order to have the same data as of the group.
The recovery process is composed by 3 stages:
  1) Apply existent transactions on group_replication_applier, if
     existent, this happens when this member already was a member of
     the group and left with transactions to apply.
  2) Fetch and apply the data the new member is missing when
     compared with the group. This happens through a asynchronous
     replication channel, named group_replication_recovery, which
     stop condition is apply all data until the moment on which the
     new member joined, marked by View_change_log_event.
     While this is ongoing, the concurrent group workload is being
     queued and will be deal with on the next stage.
  3) Apply the queued transactions during stage 2 and then switch
     the member status to ONLINE.
     The stop condition is:
       5.7: the new member will wait until the queue is empty and
            group_replication_applier applied all transactions.
       8.0: the new member will wait until the number of
            transactions queued during the 2nd stage are applied.

On this bug, at the end of the 3rd stage, the recovery process
reached its termination condition, that is, all accounted
transactions are applied.
On that *exact* moment, a new transaction is being queued. Please
remember that transactions are materialized as events on binary and
relay logs.
When the applier signals that applied all transactions, it is also
blindly searching for partial transactions, that will remove to
avoid future applier errors, which would happen if the applier did
stop at this point.
Though, this search and remove does only make sense on applier stop
cases. On execution completeness that must not be done, otherwise it
can corrupt or purge the applier relay log, which can led to data
loss.

To solve the above issue, when the applier is waiting for execution
completeness, it does not search and remove partial transactions.

+++ rapid/plugin/group_replication/src/applier.cc
@@ -786,7 +786,7 @@ Applier_module::wait_for_applier_complete_suspension(bool *abort_flag,
   {
     error= APPLIER_GTID_CHECK_TIMEOUT_ERROR; //timeout error
     while (error == APPLIER_GTID_CHECK_TIMEOUT_ERROR && !(*abort_flag))
-      error= wait_for_applier_event_execution(1); //blocking
+      error= wait_for_applier_event_execution(1, true); //blocking
   }
 
   return (error == APPLIER_RELAY_LOG_NOT_INITED);
@@ -816,7 +816,8 @@ Applier_module::is_applier_thread_waiting()
 }
 
 int
-Applier_module::wait_for_applier_event_execution(double timeout)
+Applier_module::wait_for_applier_event_execution(double timeout,
+                                                 bool check_and_purge_partial_transactions)
 {
   DBUG_ENTER("Applier_module::wait_for_applier_event_execution");
   int error= 0;
@@ -834,7 +835,8 @@ Applier_module::wait_for_applier_event_execution(double timeout)
       the applier thread will release the lock and update the applier thread
       execution position correctly and safely.
     */
-    if (((Applier_handler*)event_applier)->is_partial_transaction_on_relay_log())
+    if (check_and_purge_partial_transactions &&
+        ((Applier_handler*)event_applier)->is_partial_transaction_on_relay_log())
     {
         error= purge_applier_queue_and_restart_applier_module();
     }
